Refactor setup readiness first-run summary
