+++ src/family-directory-layout.test.ts
@@ -83,6 +83,7 @@ const EXPECTED_TOP_LEVEL_ENTRIES = {
     "run-once-turn-execution.ts",
     "setup-config-preview.ts",
     "setup-config-write.ts",
+    "setup-readiness-first-run.ts",
     "setup-readiness.ts",
     "stale-configured-bot-auto-handle.ts",
     "subprocess-safety.ts",
